Change wrappers to trait extensions
